Add files via upload
Due to changes to the SELFDESTRUCT mechanics, this level has become impassable.

More specifically: Dencun (often referred to as Cancun-Deneb) — EIP-6780 (March 2024) — "neutered" SELFDESTRUCT

EIP-6780 changed the semantics: if a SELFDESTRUCT is executed in a transaction other than the one in which the contract was created, the code/storage is NOT deleted; instead, the entire balance is simply transferred to the specified address. Only if the contract is destroyed in the same transaction in which it was created does the behavior remain the same (code/storage is deleted).